Swapped everything to double precision

+++ src/modules/Network.py
@@ -18,8 +18,8 @@ class ConnectedLayer:
         self.w = tf.clip_by_value(self.w_0, 0, float("inf"))
         self.b = tf.Variable(tf.zeros(self.outSize))
 
-        self.x = tf.placeholder(tf.float32, [None, self.inSize])
-        self.y = tf.placeholder(tf.float32, [None, self.outSize])
+        self.x = tf.placeholder(tf.float64, [None, self.inSize])
+        self.y = tf.placeholder(tf.float64, [None, self.outSize])
         
         self.z = tf.matmul(self.x, self.w) + self.b
